+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BUG FIXES:
 - Fix the bug that coverage reports are generated even if there are no valid test cases.
 - Fix the bug that other dependency resolvers are not called when error occurs in the previous dependency resolver.
 - Fix the bug that the generated resources are not in the correct order.
+- Fix the bug that `HEAD` method and other methods combination couldn't be handled correctly.
 
 ## v0.12.0
 FEATURES: